Ticker should emit Double (not UInt64)
To remain in line with the rest of the code

+++ Sources/Venice/Ticker/Ticker.swift
@@ -25,10 +25,10 @@
 import C7
 
 public final class Ticker {
-    private let internalChannel = Channel<UInt64>()
+    private let internalChannel = Channel<Double>()
     private var stopped: Bool = false
 
-    public var channel: ReceivingChannel<UInt64> {
+    public var channel: ReceivingChannel<Double> {
         return internalChannel.receivingChannel
     }
 
@@ -37,7 +37,7 @@ public final class Ticker {
             while true {
                 nap(for: period)
                 if self.stopped { break }
-                self.internalChannel.send(UInt64(now()))
+                self.internalChannel.send(now())
             }
         }
     }
